+++ tests/test_components/test_autograd.py
@@ -810,7 +810,7 @@ def test_autograd_async(use_emulated_run, structure_key, monitor_key):
     make_sim = fn_dict["sim"]
     postprocess = fn_dict["postprocess"]
 
-    task_names = {"test_a", "adjoint", "task1", "_test"}
+    task_names = {"test_a", "adjoint", "_test"}
 
     def objective(*args):
         sims = {task_name: make_sim(*args) for task_name in task_names}
@@ -910,7 +910,7 @@ def test_autograd_async_some_zero_grad(use_emulated_run, structure_key, monitor_
     make_sim = fn_dict["sim"]
     postprocess = fn_dict["postprocess"]
 
-    task_names = {"1", "2", "3", "4"}
+    task_names = {"1", "2"}
 
     def objective(*args):
         sims = {task_name: make_sim(*args) for task_name in task_names}
@@ -932,7 +932,7 @@ def test_autograd_async_all_zero_grad(use_emulated_run):
     make_sim = fn_dict["sim"]
     postprocess = fn_dict["postprocess"]
 
-    task_names = {"1", "2", "3", "4"}
+    task_names = {"1", "2"}
 
     def objective(*args):
         sims = {task_name: make_sim(*args) for task_name in task_names}